충돌 해결 07-12 19:08

+++ src/main/java/rabbit/umc/com/demo/schedule/service/ScheduleServiceImpl.java
@@ -3,7 +3,11 @@ package rabbit.umc.com.demo.schedule.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cglib.core.Local;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import rabbit.umc.com.config.BaseException;
+import rabbit.umc.com.config.BaseResponseStatus;
 import rabbit.umc.com.demo.mission.Mission;
+import rabbit.umc.com.demo.mission.MissionUsers;
 import rabbit.umc.com.demo.mission.repository.MissionRepository;
 import rabbit.umc.com.demo.schedule.domain.MissionSchedule;
 import rabbit.umc.com.demo.schedule.domain.Schedule;
@@ -25,6 +29,7 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class ScheduleServiceImpl implements ScheduleService {
 
     private final ScheduleRepository scheduleRepository;
@@ -65,23 +70,41 @@ public class ScheduleServiceImpl implements ScheduleService {
     @Override
     public ScheduleDetailRes getScheduleDetail(Long scheduleId) {
 
-        MissionSchedule missionSchedule = missionScheduleRepository.getScheduleDetail(scheduleId);
+        MissionSchedule missionSchedule = missionScheduleRepository.getMissionScheduleByScheduleId(scheduleId);
+        Schedule schedule = scheduleRepository.findScheduleById(scheduleId);
+        Mission mission = new Mission();
+        missionSchedule.setSchedule(schedule);
+        if(missionSchedule.getMission() == null){
+            missionSchedule.setMission(null);
+        }else {
+            mission = missionRepository.getMissionById(missionSchedule.getMission().getId());
+        }
+            missionSchedule.setMission(mission);
 
         return ScheduleDetailRes.setMissionSchedule(missionSchedule);
     }
 
     @Override
+    @Transactional
     public Long postSchedule(PostScheduleReq postScheduleReq, Long userId) {
         Schedule schedule = new Schedule();
-        User user = userRepository.findById(userId).get();
-
-        schedule.setSchedule(postScheduleReq,user.getId());
-
+        System.out.println("userId = " + userId);
+        User user = userRepository.getReferenceById(userId);
+        schedule.setSchedule(postScheduleReq);
+        schedule.setUser(user);
+        System.out.println("schedule = " + schedule.getId());
         // schedule에 내용들 저장
         scheduleRepository.save(schedule);
 
         MissionSchedule missionSchedule = new MissionSchedule();
-        missionSchedule.setMissionAndSchedule(postScheduleReq.getMissionId(),schedule.getId());
+        Mission mission = new Mission();
+        if(postScheduleReq.getMissionId() != null){
+            mission.setId(postScheduleReq.getMissionId());
+            missionSchedule.setMission(mission);
+        }
+
+        missionSchedule.setSchedule(schedule);
+
 
         // missionSchedule 테이블에 일정 아이디랑 미션 아이디 저장
         missionScheduleRepository.save(missionSchedule);
@@ -90,19 +113,23 @@ public class ScheduleServiceImpl implements ScheduleService {
     }
 
     @Override
-    public void deleteSchedule(Long scheduleId) {
+    @Transactional
+    public void deleteSchedule(Long scheduleId,Long userId) throws BaseException {
+        Schedule findSchedule = scheduleRepository.findScheduleById(scheduleId);
+
+        if (findSchedule==null){
+            throw new BaseException(BaseResponseStatus.FAILED_TO_SCHEDULE);
+        }
+
+        if(!findSchedule.getUser().getId().equals(userId)){
+            throw new BaseException(BaseResponseStatus.INVALID_USER_JWT);
+        }
+
+        missionScheduleRepository.deleteByScheduleId(scheduleId);
         scheduleRepository.deleteById(scheduleId);
+
     }
 
-    @Override
-    public void patchSchedule(Schedule schedule) {
-        scheduleRepository.save(schedule);
-    }
-
-    @Override
-    public Schedule findById(Long scheduleId) {
-        return scheduleRepository.findById(scheduleId).get();
-    }
 
     @Override
     public List<ScheduleListDto> getScheduleByWhen(String when) throws ParseException {
@@ -118,5 +145,22 @@ public class ScheduleServiceImpl implements ScheduleService {
 
         return scheduleHomeRes.getScheduleList();
     }
+
+    @Override
+    @Transactional
+    public void updateSchedule(PostScheduleReq patchScheduleReq, Long userId, Long scheduleId) {
+        Schedule schedule = scheduleRepository.findScheduleById(scheduleId);
+        schedule.setSchedule(patchScheduleReq);
+        MissionSchedule missionSchedule = missionScheduleRepository.findMissionScheduleByScheduleId(scheduleId);
+        if(patchScheduleReq.getMissionId() != null){
+            Mission mission = missionRepository.getReferenceById(patchScheduleReq.getMissionId());
+            missionSchedule.setMission(mission);
+        }else {
+            missionSchedule.setMission(null);
+        }
+        missionScheduleRepository.save(missionSchedule);
+        scheduleRepository.save(schedule);
+    }
+
 
 }
